Enabled cross platform input and changed to getbuttondown to be able to use the input managers Fire1 key

+++ Assets/Shine/Client/Damage.cs
@@ -3,7 +3,7 @@
 public class Damage
 {
     public DamageType Type { get; private set; }
-    public int Amount { get; private set; }
+    public float Amount { get; private set; }
 
     public Damage(DamageType type, int amount)
     {
